backend/requirements.txt
fastapi
pydantic == 2.10.4
uvicorn == 0.34.0
gunicorn == 23.0.0
asgi_correlation_id == 4.3.4
greenlet == 3.1.1
gevent == 24.11.1
python-dateutil == 2.9.0

# #Cache
# cache-fastapi == 0.0.6

#Tests
pytest == 8.3.4
pytest-mock == 3.14.0
pytest-asyncio == 0.25.0
httpx == 0.28.1

# Database
sqlmodel == 0.0.22
sqlalchemy_utils == 0.41.2
asyncpg == 0.30.0
psycopg2-binary

# Logger
loguru == 0.7.3
prometheus-fastapi-instrumentator == 7.1.0

# Code quality
ruff == 0.8.4
black == 24.10.0
flake8 == 7.1.1
isort == 5.13.2
bandit == 1.8.0
semgrep == 1.101.0
